moving generateImagesJson testhelper function in a common place
