+++ generate_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/paketo-buildpacks/packit/cargo"
 	"github.com/paketo-buildpacks/packit/v2"
 	ubinodejsextension "github.com/paketo-community/ubi-nodejs-extension"
+	"github.com/paketo-community/ubi-nodejs-extension/internal/testhelpers"
 	"github.com/paketo-community/ubi-nodejs-extension/internal/utils"
 	"github.com/paketo-community/ubi-nodejs-extension/structs"
 	"github.com/sclevine/spec"
@@ -109,7 +110,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("Specific version of node requested", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -221,7 +222,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("should return the default when node version has NOT been requested", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{true, false}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{true, false}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -306,7 +307,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("should return the higher node version when it requests for >=nodeVersion", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -380,7 +381,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("Should error on below cases of requested node", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -469,7 +470,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("Should respect the priorities and return the proper Node.js version", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -587,7 +588,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("Should have the same value as the BP_UBI_RUN_IMAGE_OVERRIDE if is not empty string", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
@@ -642,7 +643,7 @@ func testGenerate(t *testing.T, context spec.G, it spec.S) {
 
 		it("Should fallback to the run image which corresponds to the selected node version during build", func() {
 
-			imagesJsonContent := utils.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
+			imagesJsonContent := testhelpers.GenerateImagesJsonFile([]string{"16", "18"}, []bool{false, true}, false)
 			imagesJsonTmpDir = t.TempDir()
 			imagesJsonPath = filepath.Join(imagesJsonTmpDir, "images.json")
 			Expect(os.WriteFile(imagesJsonPath, []byte(imagesJsonContent), 0600)).To(Succeed())
